adds a test for createProgressBar
